+++ src/components/NewQuestionnaire/Forms/LinearScaleForm.vue
